+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * **Feat**: [336](https://github.com/SimformSolutionsPvtLtd/chatview/issues/336) Add edit message support.
 * **Feat**: [115](https://github.com/SimformSolutionsPvtLtd/chatview/issues/115) Add customizable timestamp styling for chat bubbles and improve timestamp rendering logic.
 * **Fix**: [426](https://github.com/SimformSolutionsPvtLtd/chatview/issues/426) Add auto-detection and display of voice message duration
+* **Feat**: [128](https://github.com/SimformSolutionsPvtLtd/chatview/issues/128)  Add message grouping feature to chat bubbles for improved UI
 
 ## [3.0.0]